get_score() retrieves the theta score and re-scale these at mean zero, and sd one at the population level

get_score() retrieves the theta score and re-scale these at mean zero, and sd one at the population level

Usage

get_score(model, name)

Arguments

model

an mplus object generated with techr::fit_grm2 or techr::fit_pcm

name

the name of the score included in the returned data frame

Value

data frame with id_i and a score vector with mean == 0 and sd == 1, at the population level

Examples

score_01 <- fit_grm2(
            scale_info = scales_data,
            scale_num  = 1,
            data = data_model) %>%
            get_score(., 'score')
#> Error in fit_grm2(scale_info = scales_data, scale_num = 1, data = data_model) %>%     get_score(., "score"): could not find function "%>%"